Survey page restructured
